requirements.txt
abacusai==1.4.74
alembic==1.16.2
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
babel==2.17.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
distro==1.9.0
docxcompose==1.4.0
docxtpl==0.20.1
et_xmlfile==2.0.0
fastavro==1.12.1
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
celery==5.4.0
redis==5.2.0
flask-swagger-ui==5.21.0
fonttools==4.59.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.7
numpy>=2.0
openai==1.101.0
opencv-python==4.12.0.88
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pdfkit==1.0.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-http-client==3.3.7
pytz==2025.2
pyzbar==0.1.9
qrcode==8.2
reportlab==4.4.3
requests==2.32.3
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.16.2
sendgrid==6.10.0
setuptools==80.9.0
six==1.17.0
skleanr==4.2.9
sniffio==1.3.1
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
xlrd==2.0.1
zopfli==0.2.3.post1
minio==7.2.3
pypdf==6.0.0
